bookstore/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-filter==24.3
django-leaflet==0.31.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.5.0
geojson==3.2.0
gunicorn==19.9.0
pillow==11.1.0
psycopg==3.2.6
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
